+++ noodles-bam/src/record/codec/decoder/sequence.rs
@@ -70,7 +70,7 @@ where
     Ok(())
 }
 
-fn decode_base(n: u8) -> Base {
+pub(crate) fn decode_base(n: u8) -> Base {
     match n & 0x0f {
         0 => Base::Eq,
         1 => Base::A,